enhance: support nullable group by keys (#41313)
See #36264

---------

+++ internal/core/unittest/test_expr.cpp
@@ -1408,7 +1408,7 @@ TEST_P(ExprTest, TestUnaryRangeJson) {
 
     {
         struct Testcase {
-            int64_t val;
+            double val;
             std::vector<std::string> nested_path;
         };
         std::vector<Testcase> testcases{{1.1, {"double"}},
